feat: move utils
new package for utils
move utils to the correct packages
