+++ examples/auto-book.ts
@@ -1,10 +1,12 @@
 import path from "node:path";
 
-import type { FileContentWithPath, SideEffect, ModelMessage } from "@hyv/core";
-import { Agent, createInstruction, minify, sequence, createFileWriter, writeFile } from "@hyv/core";
+import type { ModelMessage } from "@hyv/core";
+import { Agent, sequence } from "@hyv/core";
 import type { GPT4Options } from "@hyv/openai";
-import { GPTModelAdapter } from "@hyv/openai";
+import { createInstruction, GPTModelAdapter } from "@hyv/openai";
 import { Automatic1111ModelAdapter } from "@hyv/stable-diffusion";
+import { minify, createFileWriter, writeFile } from "@hyv/utils";
+import type { FileContentWithPath, SideEffect } from "@hyv/utils";
 
 /**
  * Creates a file writer for writing output files and add reading time.
