+++ packages/openai/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types.js";
+export * from "./utils.js";
 export * from "./dall-e-model-adapter.js";
 export * from "./gpt-model-adapter.js";
