+++ packages/openai/tsconfig.production.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "outDir": "dist",
     "declarationDir": "./"
-  },
+  }
 }
