+++ packages/utils/CHANGELOG.md
@@ -1,0 +1,62 @@
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.3.0](https://github.com/failfa-st/hyv/compare/v0.2.0...v0.3.0) (2023-05-06)
+
+### Features
+
+- add stable diffusion adapter
+  ([371ebed](https://github.com/failfa-st/hyv/commit/371ebed9bdc3130416047b46058fd6787c5b38a6))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.2.0](https://github.com/failfa-st/hyv/compare/v0.0.3...v0.2.0) (2023-05-04)
+
+### Features
+
+- rework the API ([#3](https://github.com/failfa-st/hyv/issues/3))
+  ([7832251](https://github.com/failfa-st/hyv/commit/78322511fa06616f5fa4de5fc89b17f337faa42d))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [0.1.0](https://github.com/failfa-st/hyv/compare/v0.0.3...v0.1.0) (2023-05-04)
+
+### Features
+
+- rework the API
+  ([db7f05e](https://github.com/failfa-st/hyv/commit/db7f05ec7264dc84849808c4f7b00cf2335aab93))
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## 0.0.3 (2023-05-04)
+
+**Note:** Version bump only for package @hyv/core
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## [0.0.2](https://github.com/failfa-st/hyv/compare/v0.0.1...v0.0.2) (2023-05-03)
+
+**Note:** Version bump only for package @hyv/core
+
+# Change Log
+
+All notable changes to this project will be documented in this file. See
+[Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+## 0.0.1 (2023-05-03)
+
+**Note:** Version bump only for package @hyv/core
