+++ packages/utils/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.production.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist",
     "declarationDir": "./"
